Completed the toroydal channel, while respecting AM overhang angle requirements
Criticalities.

The code will probably fail if the desired transition curve does not violate that requirement. Therefor in the future I will have to implement the alternative case.

The interconnection function works with local constaants, which will need to be substituted by parametric variables.

Other future steps:
close the final wall in the last toroydal section.
start to implement the input channel